Generate bindings for SCNetworkReachability.h
