+++ generate_bindings.sh
@@ -17,12 +17,14 @@ PREFERENCES_HEADER_PATH="$FRAMEWORK_PATH/SystemConfiguration.framework/Headers/S
 DYNAMIC_STORE_HEADER_PATH="$FRAMEWORK_PATH/SystemConfiguration.framework/Headers/SCDynamicStore.h"
 DYNAMIC_STORE_COPY_SPECIFIC_HEADER_PATH="$FRAMEWORK_PATH/SystemConfiguration.framework/Headers/SCDynamicStoreCopySpecific.h"
 NETWORK_CONFIGURATION_HEADER_PATH="$FRAMEWORK_PATH/SystemConfiguration.framework/Headers/SCNetworkConfiguration.h"
+NETWORK_REACHABILITY_HEADER_PATH="$FRAMEWORK_PATH/SystemConfiguration.framework/Headers/SCNetworkReachability.h"
 SCHEMA_DEFINITIONS_HEADER_PATH="$FRAMEWORK_PATH/SystemConfiguration.framework/Headers/SCSchemaDefinitions.h"
 
 PREFERENCES_BINDING_PATH="./system-configuration-sys/src/preferences.rs"
 DYNAMIC_STORE_BINDING_PATH="./system-configuration-sys/src/dynamic_store.rs"
 DYNAMIC_STORE_COPY_SPECIFIC_BINDING_PATH="./system-configuration-sys/src/dynamic_store_copy_specific.rs"
 NETWORK_CONFIGURATION_BINDING_PATH="./system-configuration-sys/src/network_configuration.rs"
+NETWORK_REACHABILITY_BINDING_PATH="./system-configuration-sys/src/network_reachability.rs"
 SCHEMA_DEFINITIONS_BINDING_PATH="./system-configuration-sys/src/schema_definitions.rs"
 
 BINDGEN_VERSION=`bindgen --version`
@@ -97,7 +99,7 @@ bindgen \
     --raw-line "use core_foundation_sys::propertylist::CFPropertyListRef;" \
     --raw-line "use core_foundation_sys::runloop::CFRunLoopRef;" \
     --raw-line "" \
-    --raw-line "use dispatch_queue_t;" \
+    --raw-line "use crate::dispatch_queue_t;" \
     --raw-line "" \
     --raw-line "pub type AuthorizationOpaqueRef = c_void;" \
     --raw-line "pub type __SCPreferences = c_void;" \
@@ -111,7 +113,6 @@ cleanup_binding $PREFERENCES_BINDING_PATH
 echo ""
 echo ""
 echo "Generating bindings for $DYNAMIC_STORE_HEADER_PATH"
-sleep 2
 
 bindgen \
     "${BINDGEN_COMMON_ARGUMENTS[@]}" \
@@ -128,7 +129,7 @@ bindgen \
     --raw-line "use core_foundation_sys::propertylist::CFPropertyListRef;" \
     --raw-line "use core_foundation_sys::runloop::CFRunLoopSourceRef;" \
     --raw-line "" \
-    --raw-line "use dispatch_queue_t;" \
+    --raw-line "use crate::dispatch_queue_t;" \
     -o $DYNAMIC_STORE_BINDING_PATH \
     $DYNAMIC_STORE_HEADER_PATH -- \
     -I$SDK_PATH/usr/include \
@@ -139,7 +140,6 @@ cleanup_binding $DYNAMIC_STORE_BINDING_PATH
 echo ""
 echo ""
 echo "Generating bindings for $DYNAMIC_STORE_COPY_SPECIFIC_HEADER_PATH"
-sleep 2
 
 bindgen \
     "${BINDGEN_COMMON_ARGUMENTS[@]}" \
@@ -150,7 +150,7 @@ bindgen \
     --blacklist-type "(__)?SCDynamicStore.*" \
     --raw-line "use core_foundation_sys::string::{CFStringEncoding, CFStringRef};" \
     --raw-line "use core_foundation_sys::dictionary::CFDictionaryRef;" \
-    --raw-line "use dynamic_store::SCDynamicStoreRef;" \
+    --raw-line "use crate::dynamic_store::SCDynamicStoreRef;" \
     --raw-line "" \
     --raw-line "use libc::c_uint;" \
     -o $DYNAMIC_STORE_COPY_SPECIFIC_BINDING_PATH \
@@ -163,13 +163,14 @@ cleanup_binding $DYNAMIC_STORE_COPY_SPECIFIC_BINDING_PATH
 echo ""
 echo ""
 echo "Generating bindings for $NETWORK_CONFIGURATION_HEADER_PATH"
-sleep 2
 
 bindgen \
     "${BINDGEN_COMMON_ARGUMENTS[@]}" \
     --whitelist-function "SCNetwork.*" \
     --whitelist-function "SCBondInterface.*" \
     --whitelist-var "kSC(NetworkInterface|BondStatus).*" \
+    --blacklist-type "SCNetworkReachability.*" \
+    --blacklist-function "SCNetworkReachability.*" \
     --blacklist-type "dispatch_queue_[ts]" \
     --blacklist-type "(__)?CF.*" \
     --blacklist-type "__SC.*" \
@@ -183,11 +184,10 @@ bindgen \
     --raw-line "use core_foundation_sys::dictionary::CFDictionaryRef;" \
     --raw-line "use core_foundation_sys::runloop::CFRunLoopRef;" \
     --raw-line "" \
-    --raw-line "use dispatch_queue_t;" \
+    --raw-line "use crate::dispatch_queue_t;" \
     --raw-line "use libc::{c_char, c_int, sockaddr, socklen_t};" \
-    --raw-line "use preferences::SCPreferencesRef;" \
+    --raw-line "use crate::preferences::SCPreferencesRef;" \
     --raw-line "" \
-    --raw-line "pub type __SCNetworkReachability = c_void;" \
     --raw-line "pub type __SCNetworkConnection = c_void;" \
     --raw-line "pub type __SCNetworkInterface = c_void;" \
     --raw-line "pub type __SCBondStatus = c_void;" \
@@ -203,13 +203,44 @@ cleanup_binding $NETWORK_CONFIGURATION_BINDING_PATH
 
 echo ""
 echo ""
+echo "Generating bindings for $NETWORK_REACHABILITY_HEADER_PATH"
+
+bindgen \
+    "${BINDGEN_COMMON_ARGUMENTS[@]}" \
+    --whitelist-function "SCNetworkReachability.*" \
+    --whitelist-type "SCNetworkReachability.*" \
+    --whitelist-var "kSCNetworkReachability.*" \
+    --blacklist-type "sockaddr" \
+    --blacklist-type "dispatch_queue_[ts]" \
+    --blacklist-type "(__)?CF.*" \
+    --blacklist-type "__SC.*" \
+    --blacklist-type "Boolean" \
+    --blacklist-type "dispatch_.*" \
+    --blacklist-type "(sockaddr|socklen_t|sa_family_t|__darwin_socklen_t|__uint.*_t)" \
+    --raw-line '#![cfg_attr(feature = "cargo-clippy", allow(clippy::unreadable_literal))]' \
+    --raw-line "use core_foundation_sys::base::{Boolean, CFAllocatorRef, CFTypeID, CFIndex};" \
+    --raw-line "use core_foundation_sys::string::CFStringRef;" \
+    --raw-line "use core_foundation_sys::runloop::CFRunLoopRef;" \
+    --raw-line "use libc::{c_char, sockaddr, c_void};" \
+    --raw-line "use crate::dispatch_queue_t;" \
+    --raw-line "pub type __SCNetworkReachability = c_void;" \
+    -o $NETWORK_REACHABILITY_BINDING_PATH \
+    $NETWORK_REACHABILITY_HEADER_PATH -- \
+    -I$SDK_PATH/usr/include \
+    -F$FRAMEWORK_PATH
+
+cleanup_binding $NETWORK_REACHABILITY_BINDING_PATH
+
+
+echo ""
+echo ""
 echo "Generating bindings for $SCHEMA_DEFINITIONS_HEADER_PATH"
-sleep 2
 
 bindgen \
     "${BINDGEN_COMMON_ARGUMENTS[@]}" \
     --whitelist-var "kSC.*" \
     --blacklist-type "(__)?CF.*" \
+    --blacklist-type "dispatch_queue_[ts]" \
     --raw-line "use core_foundation_sys::string::CFStringRef;" \
     --raw-line "" \
     -o $SCHEMA_DEFINITIONS_BINDING_PATH \
