+++ system-configuration-sys/src/dynamic_store.rs
@@ -1,8 +1,8 @@
 // automatically generated by rust-bindgen
 
 // Generated using:
-// bindgen 0.51.1
-// macOS SDK 10.15.
+// bindgen 0.54.0
+// macOS SDK 10.15.4.
 
 use core::ffi::c_void;
 use core_foundation_sys::array::CFArrayRef;
@@ -21,13 +21,20 @@ pub type SCDynamicStoreRef = *const __SCDynamicStore;
 #[repr(C)]
 pub struct SCDynamicStoreContext {
     pub version: CFIndex,
-    pub info: *mut c_void,
-    pub retain: Option<unsafe extern "C" fn(info: *const c_void) -> *const c_void>,
-    pub release: Option<unsafe extern "C" fn(info: *const c_void)>,
-    pub copyDescription: Option<unsafe extern "C" fn(info: *const c_void) -> CFStringRef>,
+    pub info: *mut ::core::ffi::c_void,
+    pub retain: Option<
+        unsafe extern "C" fn(info: *const ::core::ffi::c_void) -> *const ::core::ffi::c_void,
+    >,
+    pub release: Option<unsafe extern "C" fn(info: *const ::core::ffi::c_void)>,
+    pub copyDescription:
+        Option<unsafe extern "C" fn(info: *const ::core::ffi::c_void) -> CFStringRef>,
 }
 pub type SCDynamicStoreCallBack = Option<
-    unsafe extern "C" fn(store: SCDynamicStoreRef, changedKeys: CFArrayRef, info: *mut c_void),
+    unsafe extern "C" fn(
+        store: SCDynamicStoreRef,
+        changedKeys: CFArrayRef,
+        info: *mut ::core::ffi::c_void,
+    ),
 >;
 extern "C" {
     pub fn SCDynamicStoreGetTypeID() -> CFTypeID;
