+++ system-configuration-sys/src/dynamic_store_copy_specific.rs
@@ -1,8 +1,8 @@
 // automatically generated by rust-bindgen
 
 // Generated using:
-// bindgen 0.51.1
-// macOS SDK 10.15.
+// bindgen 0.54.0
+// macOS SDK 10.15.4.
 
 use crate::dynamic_store::SCDynamicStoreRef;
 use core_foundation_sys::dictionary::CFDictionaryRef;
