+++ system-configuration-sys/src/lib.rs
@@ -26,5 +26,6 @@ pub type dispatch_queue_t = *mut libc::c_void;
 pub mod dynamic_store;
 pub mod dynamic_store_copy_specific;
 pub mod network_configuration;
+pub mod network_reachability;
 pub mod preferences;
 pub mod schema_definitions;
