+++ system-configuration-sys/src/network_configuration.rs
@@ -1,8 +1,8 @@
 // automatically generated by rust-bindgen
 
 // Generated using:
-// bindgen 0.51.1
-// macOS SDK 10.15.
+// bindgen 0.54.0
+// macOS SDK 10.15.4.
 
 use core::ffi::c_void;
 use core_foundation_sys::array::CFArrayRef;
@@ -15,7 +15,6 @@ use crate::dispatch_queue_t;
 use crate::preferences::SCPreferencesRef;
 use libc::{c_char, c_int, sockaddr, socklen_t};
 
-pub type __SCNetworkReachability = c_void;
 pub type __SCNetworkConnection = c_void;
 pub type __SCNetworkInterface = c_void;
 pub type __SCBondStatus = c_void;
@@ -38,85 +37,24 @@ extern "C" {
 
     pub fn SCNetworkInterfaceRefreshConfiguration(ifName: CFStringRef) -> Boolean;
 }
-pub type SCNetworkReachabilityRef = *const __SCNetworkReachability;
-#[repr(C)]
-pub struct SCNetworkReachabilityContext {
-    pub version: CFIndex,
-    pub info: *mut c_void,
-    pub retain: Option<unsafe extern "C" fn(info: *const c_void) -> *const c_void>,
-    pub release: Option<unsafe extern "C" fn(info: *const c_void)>,
-    pub copyDescription: Option<unsafe extern "C" fn(info: *const c_void) -> CFStringRef>,
-}
-pub type SCNetworkReachabilityFlags = u32;
-pub type SCNetworkReachabilityCallBack = Option<
-    unsafe extern "C" fn(
-        target: SCNetworkReachabilityRef,
-        flags: SCNetworkReachabilityFlags,
-        info: *mut c_void,
-    ),
->;
-extern "C" {
-    pub fn SCNetworkReachabilityCreateWithAddress(
-        allocator: CFAllocatorRef,
-        address: *const sockaddr,
-    ) -> SCNetworkReachabilityRef;
-
-    pub fn SCNetworkReachabilityCreateWithAddressPair(
-        allocator: CFAllocatorRef,
-        localAddress: *const sockaddr,
-        remoteAddress: *const sockaddr,
-    ) -> SCNetworkReachabilityRef;
-
-    pub fn SCNetworkReachabilityCreateWithName(
-        allocator: CFAllocatorRef,
-        nodename: *const c_char,
-    ) -> SCNetworkReachabilityRef;
-
-    pub fn SCNetworkReachabilityGetTypeID() -> CFTypeID;
-
-    pub fn SCNetworkReachabilityGetFlags(
-        target: SCNetworkReachabilityRef,
-        flags: *mut SCNetworkReachabilityFlags,
-    ) -> Boolean;
-
-    pub fn SCNetworkReachabilitySetCallback(
-        target: SCNetworkReachabilityRef,
-        callout: SCNetworkReachabilityCallBack,
-        context: *mut SCNetworkReachabilityContext,
-    ) -> Boolean;
-
-    pub fn SCNetworkReachabilityScheduleWithRunLoop(
-        target: SCNetworkReachabilityRef,
-        runLoop: CFRunLoopRef,
-        runLoopMode: CFStringRef,
-    ) -> Boolean;
-
-    pub fn SCNetworkReachabilityUnscheduleFromRunLoop(
-        target: SCNetworkReachabilityRef,
-        runLoop: CFRunLoopRef,
-        runLoopMode: CFStringRef,
-    ) -> Boolean;
-
-    pub fn SCNetworkReachabilitySetDispatchQueue(
-        target: SCNetworkReachabilityRef,
-        queue: dispatch_queue_t,
-    ) -> Boolean;
-}
 pub type SCNetworkConnectionRef = *const __SCNetworkConnection;
 #[repr(C)]
 pub struct SCNetworkConnectionContext {
     pub version: CFIndex,
-    pub info: *mut c_void,
-    pub retain: Option<unsafe extern "C" fn(info: *const c_void) -> *const c_void>,
-    pub release: Option<unsafe extern "C" fn(info: *const c_void)>,
-    pub copyDescription: Option<unsafe extern "C" fn(info: *const c_void) -> CFStringRef>,
+    pub info: *mut ::core::ffi::c_void,
+    pub retain: Option<
+        unsafe extern "C" fn(info: *const ::core::ffi::c_void) -> *const ::core::ffi::c_void,
+    >,
+    pub release: Option<unsafe extern "C" fn(info: *const ::core::ffi::c_void)>,
+    pub copyDescription:
+        Option<unsafe extern "C" fn(info: *const ::core::ffi::c_void) -> CFStringRef>,
 }
 pub type SCNetworkConnectionStatus = i32;
 pub type SCNetworkConnectionCallBack = Option<
     unsafe extern "C" fn(
         connection: SCNetworkConnectionRef,
         status: SCNetworkConnectionStatus,
-        info: *mut c_void,
+        info: *mut ::core::ffi::c_void,
     ),
 >;
 extern "C" {
