+++ system-configuration-sys/src/preferences.rs
@@ -1,8 +1,8 @@
 // automatically generated by rust-bindgen
 
 // Generated using:
-// bindgen 0.51.1
-// macOS SDK 10.15.
+// bindgen 0.54.0
+// macOS SDK 10.15.4.
 
 use core::ffi::c_void;
 use core_foundation_sys::array::CFArrayRef;
@@ -23,16 +23,19 @@ pub type SCPreferencesNotification = u32;
 #[repr(C)]
 pub struct SCPreferencesContext {
     pub version: CFIndex,
-    pub info: *mut c_void,
-    pub retain: Option<unsafe extern "C" fn(info: *const c_void) -> *const c_void>,
-    pub release: Option<unsafe extern "C" fn(info: *const c_void)>,
-    pub copyDescription: Option<unsafe extern "C" fn(info: *const c_void) -> CFStringRef>,
+    pub info: *mut ::core::ffi::c_void,
+    pub retain: Option<
+        unsafe extern "C" fn(info: *const ::core::ffi::c_void) -> *const ::core::ffi::c_void,
+    >,
+    pub release: Option<unsafe extern "C" fn(info: *const ::core::ffi::c_void)>,
+    pub copyDescription:
+        Option<unsafe extern "C" fn(info: *const ::core::ffi::c_void) -> CFStringRef>,
 }
 pub type SCPreferencesCallBack = Option<
     unsafe extern "C" fn(
         prefs: SCPreferencesRef,
         notificationType: SCPreferencesNotification,
-        info: *mut c_void,
+        info: *mut ::core::ffi::c_void,
     ),
 >;
 extern "C" {
