+++ system-configuration-sys/src/schema_definitions.rs
@@ -1,8 +1,8 @@
 // automatically generated by rust-bindgen
 
 // Generated using:
-// bindgen 0.51.1
-// macOS SDK 10.15.
+// bindgen 0.54.0
+// macOS SDK 10.15.4.
 
 use core_foundation_sys::string::CFStringRef;
